Checked over packet's to what they should apparently read

+++ Terraria_Server/Messages/PlayerChestUpdate.cs
@@ -11,7 +11,7 @@ namespace Terraria_Server.Messages
 
         public void Process(int start, int length, int num, int whoAmI, byte[] readBuffer, byte bufferData)
         {
-            int inventoryIndex = BitConverter.ToInt32(readBuffer, num);
+            int inventoryIndex = (int)BitConverter.ToInt16(readBuffer, num);
             num += 2;
             int x = BitConverter.ToInt32(readBuffer, num);
             num += 4;
